5650: Flytt mapping av BekreftValg-data inn i komponenten, sletter ubrukt selector
